Fix: error handling
