+++ tests/groupchats.ts
@@ -16,10 +16,19 @@ describe('groupchats', () => {
   const groupSeed = Buffer.from(anchor.utils.bytes.utf8.encode('groupchat'))
   const inviteSeed = Buffer.from(anchor.utils.bytes.utf8.encode('invite'))
 
-  const groupId =
+  const groupIdData =
     'dhfskjdfhsdjkfhsdjkfhdsjkhdjkfdhfskjdfhsdjkfhsdjkfhdsjkhdjkfdfrt'
-  const groupHash = Buffer.from(anchor.utils.sha256.hash(groupId), 'hex')
+  const groupId = Buffer.from(groupIdData)
+  
+  const groupHash = Buffer.from(
+    anchor.utils.bytes.utf8.encode('dhfskjdfhsdjkfhsdjkfhdsjkhdjkfds'),
+  )
 
+  let name = 'dhfskjdfhsdjkfh'
+
+  let encryptionKey = 'dhfskjdfhsdjkfhsdjkfhdsjkhdjkfdhfskjdfhsdjkfhsdjkfhdsjkhdjkfdfrt'
+
+console.log(program.programId.toString())
   // Accounts for the tests.
   const group = anchor.utils.publicKey.findProgramAddressSync(
     [groupHash, groupSeed],
@@ -46,13 +55,187 @@ describe('groupchats', () => {
     program.programId,
   )
 
+  it('Creates a new group with too short name', async () => {
+    // Airdropping tokens to a payer.
+    await provider.connection.confirmTransaction(
+      await provider.connection.requestAirdrop(user1.publicKey, 10000000000),
+      'confirmed',
+    )
+    let failed = false
+    const newName = "sd"
+    try {
+      await program.rpc.create(groupHash, groupId, true, newName, encryptionKey, {
+        accounts: {
+          group: group[0],
+          invitation: inv1[0],
+          signer: user1.publicKey,
+          payer: user1.publicKey,
+          systemProgram: SystemProgram.programId,
+        },
+        signers: [user1],
+      })
+    } catch(err) {
+      const errMsg = 'The field is too short or too long'
+      assert.equal(err.toString(), errMsg)
+      failed = true
+    }
+
+    assert.ok(failed == true)
+    
+  })
+
+  it('Creates a new group with too long name', async () => {
+    // Airdropping tokens to a payer.
+    await provider.connection.confirmTransaction(
+      await provider.connection.requestAirdrop(user1.publicKey, 10000000000),
+      'confirmed',
+    )
+    let failed = false
+    const newName = "sddgijoihrgeuirg".repeat(10)
+    try {
+      await program.rpc.create(groupHash, groupId, true, newName, encryptionKey, {
+        accounts: {
+          group: group[0],
+          invitation: inv1[0],
+          signer: user1.publicKey,
+          payer: user1.publicKey,
+          systemProgram: SystemProgram.programId,
+        },
+        signers: [user1],
+      })
+    } catch(err) {
+      const errMsg = 'The field is too short or too long'
+      assert.equal(err.toString(), errMsg)
+      failed = true
+    }
+
+    assert.ok(failed == true)
+    
+  })
+
+  it('Creates a new group with too short groupid', async () => {
+    // Airdropping tokens to a payer.
+    await provider.connection.confirmTransaction(
+      await provider.connection.requestAirdrop(user1.publicKey, 10000000000),
+      'confirmed',
+    )
+    let failed = false
+    const newGroupId = Buffer.from("sd")
+    try {
+      await program.rpc.create(groupHash, newGroupId, true, name, encryptionKey, {
+        accounts: {
+          group: group[0],
+          invitation: inv1[0],
+          signer: user1.publicKey,
+          payer: user1.publicKey,
+          systemProgram: SystemProgram.programId,
+        },
+        signers: [user1],
+      })
+    } catch(err) {
+      const errMsg = 'The field is too short or too long'
+      assert.equal(err.toString(), errMsg)
+      failed = true
+    }
+
+    assert.ok(failed == true)
+    
+  })
+
+  it('Creates a new group with too long name', async () => {
+    // Airdropping tokens to a payer.
+    await provider.connection.confirmTransaction(
+      await provider.connection.requestAirdrop(user1.publicKey, 10000000000),
+      'confirmed',
+    )
+    let failed = false
+    const newGroupId = Buffer.from("sddgijoihrgeuirg".repeat(10))
+    try {
+      await program.rpc.create(groupHash, newGroupId, true, name, encryptionKey, {
+        accounts: {
+          group: group[0],
+          invitation: inv1[0],
+          signer: user1.publicKey,
+          payer: user1.publicKey,
+          systemProgram: SystemProgram.programId,
+        },
+        signers: [user1],
+      })
+    } catch(err) {
+      const errMsg = 'The field is too short or too long'
+      assert.equal(err.toString(), errMsg)
+      failed = true
+    }
+
+    assert.ok(failed == true)
+    
+  })
+
+  it('Creates a new group with wrong encryption key (63 characters)', async () => {
+    // Airdropping tokens to a payer.
+    await provider.connection.confirmTransaction(
+      await provider.connection.requestAirdrop(user1.publicKey, 10000000000),
+      'confirmed',
+    )
+    let newEncryptionKey = 'dhfskjdfhsdjkfhsdjkfhdsjkhdjkfdhfskjdfhsdjkfhsdjkfhdsjkhdjkfdfr'
+    let failed = false
+    try {
+      await program.rpc.create(groupHash, groupId, true, name, newEncryptionKey, {
+        accounts: {
+          group: group[0],
+          invitation: inv1[0],
+          signer: user1.publicKey,
+          payer: user1.publicKey,
+          systemProgram: SystemProgram.programId,
+        },
+        signers: [user1],
+      })
+    } catch(err) {
+      const errMsg = 'The field is too short or too long'
+      assert.equal(err.toString(), errMsg)
+      failed = true
+    }
+
+    assert.ok(failed == true)
+    
+  })
+
+  it('Creates a new group with wrong encryption key (65 characters)', async () => {
+    // Airdropping tokens to a payer.
+    await provider.connection.confirmTransaction(
+      await provider.connection.requestAirdrop(user1.publicKey, 10000000000),
+      'confirmed',
+    )
+    let newEncryptionKey = 'dhfskjdfhsdjkfhsdjkfhdsjkhdjkfdhfskjdfhsdjkfhsdjkfhdsjkhdjkfdfrtt'
+    let failed = false
+    try {
+      await program.rpc.create(groupHash, groupId, true, name, newEncryptionKey, {
+        accounts: {
+          group: group[0],
+          invitation: inv1[0],
+          signer: user1.publicKey,
+          payer: user1.publicKey,
+          systemProgram: SystemProgram.programId,
+        },
+        signers: [user1],
+      })
+    } catch(err) {
+      const errMsg = 'The field is too short or too long'
+      assert.equal(err.toString(), errMsg)
+      failed = true
+    }
+
+    assert.ok(failed == true)
+    
+  })
+
   it('Creates a new group', async () => {
     // Airdropping tokens to a payer.
     await provider.connection.confirmTransaction(
       await provider.connection.requestAirdrop(user1.publicKey, 10000000000),
       'confirmed',
     )
-    await program.rpc.create(groupHash, groupId, true, {
+    await program.rpc.create(groupHash, groupId, true, name, encryptionKey, {
       accounts: {
         group: group[0],
         invitation: inv1[0],
@@ -70,8 +253,58 @@ describe('groupchats', () => {
     assert.ok(invitationAccount.sender.equals(user1.publicKey))
   })
 
+  it('Admin invites new user with wrong encryption key (63 characters)', async () => {
+    let newEncryptionKey = 'dhfskjdfhsdjkfhsdjkfhdsjkhdjkfdhfskjdfhsdjkfhsdjkfhdsjkhdjkfdfr'
+    try {
+      await program.rpc.invite(groupId, user2.publicKey, newEncryptionKey, {
+        accounts: {
+          newInvitation: inv2[0],
+          group: group[0],
+          invitation: inv1[0],
+          signer: user1.publicKey,
+          payer: user1.publicKey,
+          systemProgram: SystemProgram.programId,
+        },
+        signers: [user1],
+      })
+    } catch(err) {
+      const errMsg = 'The field is too short or too long'
+      assert.equal(err.toString(), errMsg)
+    }
+
+    let groupAccount = await program.account.group.fetch(group[0])
+
+    assert.ok(groupAccount.members == 1)
+    
+  })
+
+  it('Admin invites new user with wrong encryption key (65 characters)', async () => {
+    let newEncryptionKey = 'dhfskjdfhsdjkfhsdjkfhdsjkhdjkfdhfskjdfhsdjkfhsdjkfhdsjkhdjkfdfrtt'
+    try {
+      await program.rpc.invite(groupId, user2.publicKey, newEncryptionKey, {
+        accounts: {
+          newInvitation: inv2[0],
+          group: group[0],
+          invitation: inv1[0],
+          signer: user1.publicKey,
+          payer: user1.publicKey,
+          systemProgram: SystemProgram.programId,
+        },
+        signers: [user1],
+      })
+    } catch(err) {
+      const errMsg = 'The field is too short or too long'
+      assert.equal(err.toString(), errMsg)
+    }
+
+    let groupAccount = await program.account.group.fetch(group[0])
+
+    assert.ok(groupAccount.members == 1)
+    
+  })
+
   it('Admin invites new user', async () => {
-    await program.rpc.invite(groupId, user2.publicKey, {
+    await program.rpc.invite(groupId, user2.publicKey, encryptionKey, {
       accounts: {
         newInvitation: inv2[0],
         group: group[0],
@@ -90,6 +323,7 @@ describe('groupchats', () => {
     assert.ok(invitation2Account.recipient.equals(user2.publicKey))
   })
 
+
   it('User invites new user', async () => {
     // Airdropping tokens to a payer.
     await provider.connection.confirmTransaction(
@@ -97,7 +331,7 @@ describe('groupchats', () => {
       'confirmed',
     )
 
-    await program.rpc.invite(groupId, user3.publicKey, {
+    await program.rpc.invite(groupId, user3.publicKey, encryptionKey, {
       accounts: {
         newInvitation: inv3[0],
         group: group[0],
@@ -116,11 +350,10 @@ describe('groupchats', () => {
     assert.ok(invitation2Account.recipient.equals(user3.publicKey))
   })
 
-  it('Admin modifies group settings', async () => {
-    await program.rpc.modify(false, {
+  it('Admin modifies group settings for open invites', async () => {
+    await program.rpc.modifyOpenIvites(false, {
       accounts: {
         group: group[0],
-        successor: inv1[0],
         admin: user1.publicKey,
       },
       signers: [user1],
@@ -130,12 +363,81 @@ describe('groupchats', () => {
 
     assert.ok(groupAccount.members == 3)
     assert.ok(!groupAccount.openInvites)
+    assert.ok(groupAccount.name == name)
     assert.ok(groupAccount.admin.equals(user1.publicKey))
+  })
+
+  it('Admin modifies group settings for changing name', async () => {
+    const newName = 'sdkjcvhsolikj'
+    await program.rpc.modifyName(newName, {
+      accounts: {
+        group: group[0],
+        admin: user1.publicKey,
+      },
+      signers: [user1],
+    })
+
+    let groupAccount = await program.account.group.fetch(group[0])
+
+    assert.ok(groupAccount.members == 3)
+    assert.ok(!groupAccount.openInvites)
+    assert.ok(groupAccount.name == newName)
+    assert.ok(groupAccount.admin.equals(user1.publicKey))
+
+    name = newName;
+  })
+
+  it('Admin modifies group settings for changing name with too short field', async () => {
+    const newName = 'sd'
+    try {
+      await program.rpc.modifyName(newName, {
+        accounts: {
+          group: group[0],
+          admin: user1.publicKey,
+        },
+        signers: [user1],
+      })
+    } catch(err) {
+      const errMsg = 'The field is too short or too long'
+      assert.equal(err.toString(), errMsg)
+    }
+
+    let groupAccount = await program.account.group.fetch(group[0])
+
+    assert.ok(groupAccount.members == 3)
+    assert.ok(!groupAccount.openInvites)
+    assert.ok(groupAccount.name == name)
+    assert.ok(groupAccount.admin.equals(user1.publicKey))
+
+  })
+
+  it('Admin modifies group settings for changing name with too long field', async () => {
+    const newName = 'sddsoksopivjoifdvpos'.repeat(10)
+    try {
+      await program.rpc.modifyName(newName, {
+        accounts: {
+          group: group[0],
+          admin: user1.publicKey,
+        },
+        signers: [user1],
+      })
+    } catch(err) {
+      const errMsg = 'The field is too short or too long'
+      assert.equal(err.toString(), errMsg)
+    }
+
+    let groupAccount = await program.account.group.fetch(group[0])
+
+    assert.ok(groupAccount.members == 3)
+    assert.ok(!groupAccount.openInvites)
+    assert.ok(groupAccount.name == name)
+    assert.ok(groupAccount.admin.equals(user1.publicKey))
+    
   })
 
   it('User now cannot invite new user', async () => {
     try {
-      await program.rpc.invite(groupId, user4.publicKey, {
+      await program.rpc.invite(groupId, user4.publicKey, encryptionKey, {
         accounts: {
           newInvitation: inv4[0],
           group: group[0],
@@ -156,12 +458,11 @@ describe('groupchats', () => {
     assert.ok(groupAccount.members == 3)
   })
 
-  it('User cannot modify group settings', async () => {
+  it('User cannot modify group settings for open invites', async () => {
     try {
-      await program.rpc.modify(true, {
+      await program.rpc.modifyOpenIvites(true, {
         accounts: {
           group: group[0],
-          successor: inv2[0],
           admin: user2.publicKey,
         },
         signers: [user2],
@@ -174,6 +475,29 @@ describe('groupchats', () => {
 
     let groupAccount = await program.account.group.fetch(group[0])
     assert.ok(!groupAccount.openInvites)
+    assert.ok(groupAccount.name == name)
+    assert.ok(groupAccount.admin.equals(user1.publicKey))
+  })
+
+  it('User cannot modify group settings for changing name', async () => {
+    const newName = 'kjlnklfvokofjg'
+    try {
+      await program.rpc.modifyName(newName, {
+        accounts: {
+          group: group[0],
+          admin: user2.publicKey,
+        },
+        signers: [user2],
+      })
+      assert.ok(false)
+    } catch (err) {
+      const errMsg = 'User cannot perform this action'
+      assert.equal(err.toString(), errMsg)
+    }
+
+    let groupAccount = await program.account.group.fetch(group[0])
+    assert.ok(!groupAccount.openInvites)
+    assert.ok(groupAccount.name == name)
     assert.ok(groupAccount.admin.equals(user1.publicKey))
   })
 
@@ -239,7 +563,7 @@ describe('groupchats', () => {
   })
 
   it('Admin invites old user back', async () => {
-    await program.rpc.invite(groupId, user2.publicKey, {
+    await program.rpc.invite(groupId, user2.publicKey, encryptionKey, {
       accounts: {
         newInvitation: inv2[0],
         group: group[0],
@@ -320,7 +644,7 @@ describe('groupchats', () => {
   })
 
   it('New Admin modifies group settings, changing admin', async () => {
-    await program.rpc.modify(true, {
+    await program.rpc.modifySuccessor({
       accounts: {
         group: group[0],
         successor: inv3[0],
@@ -332,7 +656,27 @@ describe('groupchats', () => {
     let groupAccount = await program.account.group.fetch(group[0])
 
     assert.ok(groupAccount.members == 2)
-    assert.ok(groupAccount.openInvites)
+    assert.ok(groupAccount.admin.equals(user3.publicKey))
+  })
+
+  it('User modifies group settings, changing admin', async () => {
+    try {
+      await program.rpc.modifySuccessor({
+        accounts: {
+          group: group[0],
+          successor: inv3[0],
+          admin: user1.publicKey,
+        },
+        signers: [user1],
+      })
+    } catch(err) {
+      const errMsg = 'User cannot perform this action'
+      assert.equal(err.toString(), errMsg)
+    }
+
+    let groupAccount = await program.account.group.fetch(group[0])
+
+    assert.ok(groupAccount.members == 2)
     assert.ok(groupAccount.admin.equals(user3.publicKey))
   })
